Implement async logger
The async logger won't slow down LFXE anymore when it's in debug log
level. It has its own thread that gets signaled when there's something
to log. This does mean that the logger may run behind the actual things
that are happening. But it will get written to the file eventually.